+++ tests/test_podvector.py
@@ -9,27 +9,19 @@ def test_podvector_init():
     print(podv.__array_interface__)
     # podv[0] = 1
     # podv[2] = 3
-    print(podv.size())
     assert(podv.size() == 0)
     podv.push_back(1)
     podv.push_back(2)
-    print(podv.size())
     assert(podv.size() == 2 and podv[1] == 2)
-    print(podv)
     podv.pop_back()
     assert(podv.size() == 1)
-    print(podv)
     podv.push_back(2.14)
-    print('Is PODVector empty?', podv.empty())
     assert(not podv.empty())
     podv.push_back(3.1)
     podv[2] = 5
-    print(podv)
     assert(podv.size() == 3 and podv[2] == 5)
     podv.clear()
     assert(podv.size() == 0)
-    print(podv)
-    print('Is PODVector empty now?', podv.empty())
     assert(podv.empty())
 
 def test_array_interface():
@@ -38,9 +30,14 @@ def test_array_interface():
     podv.push_back(2)
     podv.push_back(1)
     podv.push_back(5)
-    myarr = np.array([1,2,1,5])
-    arr = np.array(podv)
+    arr = np.array(podv, copy=False)
     print(arr)
-    # print(myarr.__array_interface__)
-    # pv = amrex.PODVector_int(myarr)
-    # print(pv.__array_interface__)
+
+    # podv[2] = 3
+    arr[2] = 3
+    print(arr)
+    print(podv)
+    assert(arr[2]==podv[2]==3)
+
+    podv[1] = 5
+    assert(arr[1]==podv[1]==5)
